backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
python-multipart>=0.0.9
pydantic>=2.7
numpy>=1.26
scikit-learn>=1.4
pymupdf
pdfplumber
sentence-transformers
umap-learn
